Remove attempted re-export of `Prim.Row.Cons`
It is currently possible to attempt to re-export both a class and a type
of the same name from different modules.
It is not currently possible to define both a class and a type of the
same name in the same module.
What actually happens is that the order of exports defines which things
are really exported.

In the case of `Type.Row`, the re-export of `Prim.Row` comes first,
and the re-export of `Prim.RowList` comes after it.
This ordering means that the `Prim.Row.Cons` class is not re-exported,
but the `Prim.RowList.Cons` type is re-exported.

From the perspective of anyone importing `Type.Row`,
we've always only re-exported `Prim.RowList.Cons`.
We make that explicit by removing `Prim.Row.Cons` from the export list.